requirements.txt
africastalking==2.0.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
audioop-lts==0.2.2
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
CacheControl==0.14.4
cbor2==5.8.0
celery==5.3.4
certifi==2026.2.25
cffi==2.0.0
channels==4.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.5
daphne==4.0.0
dj-database-url==2.2.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.20.0
firebase-admin==6.2.0
frozenlist==1.8.0
fsspec==2025.12.0
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==25.1.0
httplib2==0.31.2
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
Jinja2==3.1.6
kombu==5.6.2
llvmlite==0.46.0
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
openai-whisper==20250625
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydub==0.25.1
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2026.1.post1
PyYAML==6.0.3
redis==5.0.1
regex==2026.2.28
requests==2.32.5
responses==0.26.0
rsa==4.9.1
schema==0.7.8
service-identity==24.2.0
setuptools==82.0.0
six==1.17.0
SpeechRecognition==3.10.0
sqlparse==0.5.5
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0+cpu
tqdm==4.67.3
triton==3.6.0
twilio==8.10.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
wheel==0.46.3
whitenoise==6.12.0
yarl==1.23.0
zope.interface==8.2
